moving SortedAdventure type to their own file
