+++ website/frontend_types.py
@@ -103,15 +103,3 @@ class Adventure:
             start_code='',  # Teacher adventures don't seem to have this
             text=row['content'],
             is_teacher_adventure=True)
-
-@dataclass
-class SortedAdventure:
-    short_name: str
-    long_name: str
-    is_teacher_adventure: bool
-    is_command_adventure: bool
-
-    # Meant to be used when testing membership against
-    # a list of strings
-    def __hash__(self) -> int:
-        return hash(self.short_name)
